Maintain an inverted index.
An inverted index tells us for each term, which documents contain
this term.  We also include the term frequency (number of times
the term occurs in the document), and the document frequency
(number of documents that have this term).

The `vocabularies` table contains the list of available terms
and their document frequencies.

The `terms` table contains for each term the documents it occurs
in, and the term frequency in that document.